module/command architecture
