+++ 3520/2/commands/checkin.rb
@@ -1,0 +1,5 @@
+module CheckinCommand
+  def checkin!
+    respond_with :message, text: 'Send me a selfire.'
+  end
+end
