+++ 3520/2/commands/logout.rb
@@ -1,0 +1,10 @@
+module LogoutCommand
+  def logout!
+    if registered?
+      respond_with :message, text: "Goodbye, #{session[:number]}!"
+      session.destroy
+    else
+      respond_with :message, text: 'Nice try!'
+    end
+  end
+end
